feat: Split input from output scanners
