+++ src/Andy.Guard/Andy.Guard.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.ML.Tokenizers" Version="2.0.0-preview.25373.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="OutputScanners/" />
+  </ItemGroup>
+
 </Project>
