+++ tests/Andy.Guard.Tests/Tokenizer/DebertaTokenizerTests.cs
@@ -1,0 +1,23 @@
+using System;
+
+namespace Andy.Guard.Tests.Tokenizer;
+
+public class DebertaTokenizerTests
+{
+    // TODO (HARD but Critical) Testing for parity
+
+    // To ensure parity with HuggingFace:
+    // 1. Tokenize a few sentences with HuggingFace in Python and save (input_ids, attention_mask).
+    // 2. Run the same through this C# code.
+    // 3. Compare arrays element-wise — they should match exactly.
+
+    // If you see differences:
+    // • Double-check the special-token IDs.
+    // • Make sure case is preserved (don’t lowercase).
+    //     • Verify max length and truncation strategy are the same.
+    public void GivenIdenticalSentences_Encode_ReturnsIdenticalEncodingThanTheHuggingFaceImplementation()
+    {
+        // Placeholder test to ensure the test project builds and runs.
+        Assert.True(true);
+    }
+}
